feat/#3 : Components _ text input(thin) 추가
